+++ src/parser/enrichers/monster/GiantInsect/_module.mjs
@@ -1,0 +1,2 @@
+export { default as PoisonJab } from "./PoisonJab.mjs";
+export { default as VenomousSpew } from "./VenomousSpew.mjs";