+++ src/parser/enrichers/monster/_module.mjs
@@ -14,3 +14,4 @@ export * as Empyrean from "./Empyrean/_module.mjs";
 export * as SpiritualWeapon from "./SpiritualWeapon/_module.mjs";
 export * as FaithfulHound from "./FaithfulHound/_module.mjs";
 export * as GuardianOfFaith from "./GuardianOfFaith/_module.mjs";
+export * as GiantInsect from "./GiantInsect/_module.mjs";